Fix out of memory error during build pyramid operation on worldwide raster

+++ mrgeo-core/src/main/java/org/mrgeo/data/tile/MrsTileReader.java
@@ -129,7 +129,7 @@ public abstract class MrsTileReader<T>
    */
   public abstract void close();
 
-  public abstract int calculateTileCount();
+  public abstract long calculateTileCount();
 
   /**
    * Varify if an item exists in the data.
